Placed user controllers in seperate files
